+++ components/persistor/HISTORY.md
@@ -1,3 +1,11 @@
+## 9.0.2
+* package-lock bump to pick up latest supertype version.
+* added test case for testing _ids during object cloning.
+## 9.0.1
+* Added executionPromise to knex commit and rollback
+* Explicitly return promise from knex transaction
+## 9.0.0
+* update persistor to use client passed logger.
 ## 8.0.2
 * bump nconf version due to security vulnerability.
 ## 8.0.1