+++ components/semotus/HISTORY.md
@@ -1,3 +1,7 @@
+## 7.0.1
+* package-lock bump to pick up latest supertype version.
+## 7.0.0
+* update semotus to use client passed logger.
 ## 6.0.1
 * bump nconf version due to security vulnerability.
 ## 6.0.0